Adding validation status to product details downloads

+++ view_models/template_funcs.go
@@ -100,8 +100,8 @@ func formatBytes(b int) string {
 		float64(b)/float64(div), "kMGTPE"[exp])
 }
 
-func unixDateFormat(d int) string {
-	return time.Unix(int64(d), 0).Format("Jan 2, 2006")
+func unixDateFormat(d int64) string {
+	return time.Unix(d, 0).Format("Jan 2, 2006")
 }
 
 func unixDateTimeFormat(d int64) string {
